+++ .gitignore
@@ -13,3 +13,4 @@ docker-compose.override.yml
 **/yarn.lock
 client/src/components/Input/FormInput ORIGINAL.js
 client/src/components/Positions/PositionApplicationForm - Original.js
+client/src/components/Positions/PositionApplicationForm - Original.js
